# app/static/spec/swagger.yaml
---
openapi: 3.0.0
info:
  description: Peer-to-Peer Energy Trading in Electrical Distribution Networks
  version: 1.0.0
  type: Open-ElecBay
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "/"
tags:
  - name: Market
    description: Support P2P bidding
paths:
  "/request":
    get:
      tags:
        - Market Request
      summary: Returns marketRequests
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/components/schemas/marketRequests"
    post:
      tags:
        - Market Request
      summary: Create a new market request system
      requestBody:
        description: Market Request Post Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/marketRequestPostBody"
      produces:
        - application/json
      responses:
        "201":
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        "400":
          description: Failed. Bad post data.
  "/request/submit/{id}":
    post:
      tags:
        - Market Request
      summary: Submits the marketRequests
      parameters:
        - in: path
          name: id
          required: true
          description: Market Request id
          schema:
            "$ref": "#/components/schemas/id"
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/components/schemas/marketRequest"
        "400":
          description: Failed. Misunderstood Request.
        "404":
          description: Failed. Market request not found.

  "/request/{id}":
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the request that we want to match
        type: string
    get:
      tags:
        - Market Request
      summary: Get market request with given ID
      parameters:
        - in: path
          name: id
          required: true
          description: Market Request id
          schema:
            "$ref": "#/components/schemas/id"
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/components/schemas/marketRequest"
        "400":
          description: Failed. Misunderstood Request.
        "404":
          description: Failed. Market request not found.
    put:
      summary: edit a market request by ID
      tags:
        - Market Request
      parameters:
        - in: path
          name: id
          required: true
          description: Market Request id
          schema:
            "$ref": "#/components/schemas/id"
      requestBody:
        description: Market Request Object
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/marketRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            "$ref": "#/components/schemas/marketRequest"
        "400":
          description: Failed. Bad post data.
    delete:
      summary: Delete Market Request by ID
      tags:
        - Market Request
      parameters:
        - in: path
          name: id
          required: true
          description: Market Request Id
          schema:
            "$ref": "#/components/schemas/id"
      responses:
        "204":
          description: OK
          schema:
            "$ref": "#/components/schemas/id"
        "400":
          description: Failed. Misunderstood Request.
        "404":
          description: Failed. Market Request not found.
components:
  schemas:
    id:
      properties:
        uuid:
          type: string
    marketRequestPostBody:
      type: object
      properties:
        type:
          type: string
          format: string
        message:
          type: string
          format: string
    marketRequest:
      type: object
      properties:
        type:
          type: string
          format: string
        message:
          type: string
          format: string
        timestamp:
          type: string
          format: number
    marketRequests:
      type: object
      properties:
        marketRequest:
          type: object
          additionalProperties:
            "$ref": "#/components/schemas/marketRequest"